+++ CHANGELOG.md
@@ -3,6 +3,14 @@ Change Log
 
 All notable changes to this project will be documented in this file.
 
+[Unreleased](https://github.com/FreddieChopin/bleeding-edge-toolchain/compare/221119...HEAD)
+--------------------------------------------------------------------------------------------
+
+### Fixed
+
+- Added temporary fix for building GDB 12.1 documentation with Texinfo 7
+([upstream patch](https://git.sr.ht/~sourceware/binutils-gdb/commit/7bd836d5d90353a2de192fd4711a20b4520246b7)).
+
 [221119](https://github.com/FreddieChopin/bleeding-edge-toolchain/compare/221106...221119)
 ------------------------------------------------------------------------------------------